"Changed price type from double to int in Recommendation model and updated related files."

+++ lib/main/models/recommendation.dart
@@ -1,5 +1,3 @@
-import 'dart:ffi';
-
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'recommendation.freezed.dart';
@@ -12,7 +10,7 @@ class Recommendation with _$Recommendation {
         @JsonKey(name: "title")
         String? title,
         @JsonKey(name: "price")
-        double? price,
+        int? price,
         @JsonKey(name: "desc")
         String? desc,
         @JsonKey(name: "rating")
